chore: reorganize src directory
Moved GeoTIFFTileSource class to it's own top level file that gets imported into main.js
